content-service/etl/requirements.txt
certifi==2022.9.24
elastic-transport==8.4.0
elasticsearch==7.10.1
psycopg2==2.9.5
pydantic==1.10.2
python-dotenv==0.21.0
typing_extensions==4.4.0
urllib3==1.26.12
